Shared 'https' handling for display, 3d viewer, databrowser, ..
.. and anything else that builds on ResourceParser to open https://
streams
